elasticsearch/requirements.txt
elasticsearch
kafka-python
black
python-dotenv
pylint
